refactor: modularize useProgramStore into domain slices and cleanup legacy V2 logic (fixes #38)
