+++ _project-system/knowledge-base/history/handoffs/handoff-2026-01-26T14-00.md
@@ -1,7 +1,7 @@
 # Handoff - 2026-01-26T14:00
 
 ## Session Overview
-In this session, we successfully evolved "Master Plan" into a dual-environment platform by implementing and hardening the **Sandbox (v3_dev)** environment. This allows for risk-free testing of new features (like program deletion and editing) without affecting production logs.
+In this session, we successfully evolved "Master Plan" into a dual-environment platform by implementing and hardening the **Sandbox (v3_dev)** environment. The session concluded with a successful UAT, merge to `main`, and closure of tracking issues.
 
 ## Critical Changes
 
@@ -28,13 +28,14 @@ In this session, we successfully evolved "Master Plan" into a dual-environment p
 - **Active Environment:** Support for dynamic schema switching via HubApp.
 - **Schema Parity:** `v3_dev` is now a perfect mirror of `v3` with functioning relationships.
 - **Version:** Platform bumped to **v1.6.0**.
+- **Source Control:** Feature branch `task-35-expose-dev-schema` merged into `main` and deleted.
+- **Project Board:** GHI-34 and GHI-35 closed and verified.
 
 ## Developer Protocol Updates (Lessons Learned)
 - **Rule - Schema Cloning:** Cloned schemas *must* have Foreign Keys manually re-established within the new schema scope; PostgreSQL `LIKE` does not redirect them.
 - **Rule - API Exposure:** New schemas require a Management API `PATCH` to `db_schemas` for PostgREST visibility.
 - **Rule - Header HUDs:** Use absolute positioning for environment markers to prevent pushing core navigation/tactical elements (Timers, Abandon buttons) out of alignment.
 
-## Pending Tasks
+## Next Steps
 - [ ] Implement structural sync scripts (automated SQL migrations) for schema parity.
 - [ ] Implement on-demand data sync (Prod -> Dev cloning) via UI trigger.
-- [ ] Merge `task-35-expose-dev-schema` into main.
