+++ _project-system/knowledge-base/history/handoffs/handoff-2026-01-26T18-00.md
@@ -1,0 +1,30 @@
+# Handoff - 2026-01-26T18:00
+
+## Session Overview
+In this session, we transitioned from a monolithic "God Object" store to a **Modular Domain-Driven Slice Architecture**. This establishes a best-in-class foundation for future scalability while maintaining 100% of the relational hardening logic implemented in v1.7.0.
+
+## Critical Changes
+
+### 1. Store Modularization (Zustand Slice Pattern)
+- **`useProgramStore.js`**: Now a unified entry point composed of specialized slices.
+- **`slices/uiSlice.js`**: Handles navigation context and schema resolution.
+- **`slices/programSlice.js`**: Manages program CRUD, surgical builder updates, and library previews.
+- **`slices/sessionSlice.js`**: Controls the active workout focus engine and set logging.
+- **`slices/historySlice.js`**: Manages the performance vault, volume analytics, and AI exports.
+
+### 2. Service Layer Abstraction
+- **`services/database.service.js`**: Centralized all schema-aware queries into a single repository. This eliminates the risk of data leakage by ensuring all database interaction happens through a unified, schema-validated interface.
+
+### 3. State Preservation & Hardening
+- **Optimized Persistence**: Refactored the `persist` middleware to only cache essential state (Session recovery, Navigation context), reducing local storage overhead.
+- **Logic Integrity**: Verified that the **Snapshot Strategy** and **Surgical Save** logic were accurately ported to the modular slices.
+
+## Current State
+- **Architecture:** Modular, Encapsulated, and Scalable.
+- **Database:** Fully hardened with Snapshot columns and unique session constraints.
+- **Stability:** High. Production build verified successfully.
+
+## Next Steps
+- [ ] **Structural Sync Script:** Implement the automated migration tool to keep Prod/Sandbox schemas aligned.
+- [ ] **Data Cloning Utility:** Implement the "Prod -> Dev" UI trigger for easy environment mirroring.
+- [ ] **Legacy Cleanup:** Mark `useTrainingStore.js` as deprecated and move V2 logic to a legacy adapter if needed.
