+++ package.json
@@ -18,6 +18,7 @@
     "@tailwindcss/vite": "^4.1.18",
     "framer-motion": "^12.28.1",
     "lucide-react": "^0.309.0",
+    "pg": "^8.17.2",
     "react": "^18.2.0",
     "react-day-picker": "^9.13.0",
     "react-dom": "^18.2.0",
